Moved tests to es2015
