+++ package.json
@@ -32,6 +32,7 @@
     "jasmine-node": "^1.14.5",
     "jsdom": "^3.1.2",
     "react": "^0.14.0-beta3",
+    "react-addons-test-utils": "^0.14.0-beta3",
     "react-dom": "^0.14.0-beta3",
     "webpack": "^1.7.3",
     "webpack-dev-server": "^1.7.0"
