+++ specs/utils/immediate.js
@@ -1,0 +1,3 @@
+export default function (fn) {
+  setTimeout(fn, 0);
+}
